requirements.txt
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==2.0.6
requests==2.9.1
djangorestframework==3.5
djangorestframework-jwt==1.8.0
coreapi==2.1.1
pylint==1.7.2
pylint-django==0.7.2
autopep8==1.3.2
imagekit==1.0.6
pillow==4.2.1
django-extra-fields==0.9